refactor: Extract makeVStorage to separate vstorage.js file
